[level 0] Title: 나머지 구하기, Time: 0.03 ms, Memory: 77.5 MB -BaekjoonHub

+++ 프로그래머스/0/120810. 나머지 구하기/나머지 구하기.java
@@ -1,6 +1,6 @@
 class Solution {
-    public float solution(float num1, float num2) {
-        float answer = num1 % num2;
+    public int solution(int num1, int num2) {
+        int answer = num1 % num2;
         
         return answer;
     }
